requirements.txt
boto3==1.28.83
botocore==1.31.83
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
